blockchain: No context dep checks for orphans.
The sanity functions, as their comments indicate, are supposed to be
context free since they are used on blocks and transactions regardless
of their position within the chain.

Unfortunately, the decentralized treasury work ignored that and
introduced context to them.  The only current side effect is that blocks
created with the new rules, which are naturally orphans to syncing nodes
which do not yet have all of those parents, fail the sanity checks which
results in stalling the sync process until timeouts occur.

Ultimately all of those context dependent checks need to be moved into
the appropriate functions, because future work also depends on those
requirements.

However, since that is a non-trivial task, because each one has to be
fully examined to ensure there are no later checks which have implicit
assumption about them already having been done, this implements a stop
gap solution to work around the sync issues by only doing the context
free sanity checks prior to orphan determination and the context
dependent ones just after that.