blockchain: Revert fast add reversal.
The decentralized treasury commit modified the fast add checks to exit
early in a couple of functions instead of executing the code inside of
the branch.

This reverts that change because even though there are not currently any
checks which are done that are not part of that branch, it is highly
likely there will be in the future and thus it was very intentionally
not short circuited in that way.