blockchain: Less order dependent full blocks tests.
This modifies some of the full block tests to make them accurate
regardless of the order the rules are tested in.

In particular, it updates the fresh stake value in the header of the
tests to not account for transactions that are not valid tickets and
also to modify transactions that are duplicated to ensure they have
unique hashes.