blockchain: Move context free block sanity checks.
This moves all block sanity checks which are possible to check without
context into the context free function versus the contextual one.

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.