blockchain: Rework contextual tx checks.
This reworks the existing checks categorized as contextual transaction
sanity checks to more closely mirror the way block checks are done such
that there is a context-free sanity func and a separate contextual func
as well as update several things along the way such as improving and/or
correcting the error messages.

The following is a high level overview of the changes:

- Rename checkTransactionSanityContextual to checkTransactionContext
- Move invocation of said transaction checks to checkBlockContext
- Avoid using fmt.Sprintf for static strings that take no arguments
- Update the text of several errors to be more consistent with the rest
- Add the required length for treasury spend scripts to error message
- Add the input index to error message for unexpected null outpoints
- Correct and/or improve several comments both in the code and the tests
- Make wrapping consistent
- Update expected error in a few tests to match the new ordering

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.