blockchain: Move {coin,trsy}base contextual checks.
This moves the checks related to coinbases and treasurybases from the
sanity and positional block checks to the context-dependent block checks
func where they belong since they depend on the result of a vote and
also cleans up a few things with the new treasury code along the way.

In particular, the following is a high level overview of the changes:
- Move coinbase and treasurybase to checkBlockContext
- Remove incorrectly named ErrFirstTxNotOpReturn
- Rework the coinbase and treasurybase height checks:
  - Split coinbase and treasurybase height checks into separate funcs
  - Make checkCoinbaseUniqueHeight choose the correct output according
    to the result of the treasury agenda
  - Base minimum required outputs on the output index instead of hard
    coding it so it is correct for both cases
  - Correct a bunch of errors to include the actual output index since
    it varies depending on whether or not the agenda is active
- Correct and/or improve several comments both in the code and the tests
- Make wrapping consistent
- Update expected error in a few tests to match the new ordering

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.