blockchain: Move staketx-related contextual checks.
This moves the checks related to stake transactions that depend on the
result of a vote from the sanity block checks to the context-dependent
block checks func and also cleans up a few things along the way.

In particular, the following is a high level overview of the changes:
- Move context-dependent staketx checks to checkBlockContext
- Rename some variables for consistency
- Include number of expected stake txns in error for unexpected stake
  txns prior to stake validation height
- Make wrapping consistent
- Update expected error in a few tests to match the new ordering

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.