blockchain: Make CheckBlockSanity context free.
This removes the now unused checkBlockSanityContextual since all the
checks that rely on context are now in the proper context-based
function and restores the context-free nature of checkBlockSanity and
its exported variant.

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.