blockchain: Move sigop-related contextual checks.
This moves the checks related to counting signature operations from the
sanity block checks to the context-dependent block checks func since
they depend on the result of a vote.

While moving the code, it also makes it a bit safer against future
potential changes by creating a separate slice rather than appending
directly to the regular transactions slice since that could potentially
end up modifying the backing array depending on how it was created.

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.