blockchain: Move contextual treasury spend checks.
This moves the checks that enforce treasury spends can only be present
in blocks that fall on a TVI after the first one from the positional
block checks to the context-dependent block checks func since they
depend on the result of a vote.

While here, it also improves the error messages, optimizes the checks a
bit, and updates the later checks to rely on them already having been
done as opposed to checking again.

In particular, the following is a high level overview of the changes:
- Move context-dependent treasury spend checks to checkBlockContext
- Rename some variables for consistency
- Keep track of treasury spend transactions while counting them to avoid
  another loop over all the stake transactions and associated duplicate
  type determination
- Include transaction hash in errors
- Make the error when a treasury spend is in a block that does not fall
  on a TVI include the next TVI block height instead of the number of
  blocks in an interval
- Make wrapping consistent

All sanity and positional checks have been restored to their expected
context-based semantics as of this commit.