blockchain: Context free CheckTransactionSanity.
This renames checkTransactionSanityContextFree back to
checkTransactionSanity and restores the context-free nature of of the
exported CheckTransactionSanity variant.

It also introduces a new CheckTransaction function which includes the
context free checks as well as the contextual checks that depend on the
result of a vote and updates mempool to use that variant since it has
access to voting results and therefore should include the extra checks
to prevent invalid transactions as early as possible prior to doing more
expensive checks.

This is part of an overall effort to ultimately move all checks that
require full contextual information into the proper location since, as
the comments call out, sanity functions are supposed to be entirely
context free and positional functions are not supposed to have any
checks that rely on having the full block data of all ancestors
available.

These restrictions are important for planned future work.