blockchain: Convert to LRU block cache.
This changes the main chain block cache to a generic LRU cache for
blocks that have been recently seen.  This is useful because it is less
dependent on the order the block data is seen and it handles its own
pruning with LRU semantics.

This currently doesn't make much difference because the block data is
processed in the same order it received, but that will not always be the
case, so it is beneficial to have cache semantics which do not rely on
the exact order of received blocks.