blockmanager: Rework chain current logic.
This reworks that logic that determines whether or not the block manager
believes the chain is current (synced) to use a flag that is protected
by a separate mutex and is updated on the fly versus needing to go
through a channel and using the current sync peer.

While the primary objective is to decouple the block manager from the
server so it can be split out into a separate internal netsync package,
this also has the benefit of much faster state querying and allowing
looser coupling of block processing without the potential of deadlocks.