progresslog: Make block progress log internal.
This splits the block progress logging used by the block manager and
dump blockchain option to a separate internal package named progresslog
and updates its code to be more consistent with the rest of the code
base along the way.

This is a part of the overall effort to decouple the block manager from
the server so it can be split out into a separate internal netsync
package.