netsync: Split blockmanager into separate package.
This does the minimum work necessary to refactor the block manager into
its own internal package named netsync.  The main motivation is that
separating this code into its own package will improve its testability
and more cleanly split the details related to syncing with the network
from consensus validation.

The API will certainly need some additional cleanup and changes to make
it more usable outside of the specific circumstances it currently
supports, however it is better to do that in future commits in order to
keep the changeset as small as possible during this refactor.

Overview of the major changes:

- Create the new package
- Move blockmanager.go -> internal/netsync/manager.go
- Update logging to use the new netsync package logger
- Rename blockManagerConfig to Config (now netsync.Config)
- Move peerNotifier to netsync/interface.go (now netsync.PeerNotifier)
- Move syncManager to netsync/interface.go (now netsync.SyncManager)
- Rename newBlockManager to New (now netsync.New)
- Update all references to the block manager to use the package
- Add package logger
- Initialize new package logger with sync logger