netsync: Improve is current detection.
This modifies the logic that deals with determining if the sync manager
believes the chain is current (aka fully synced) to ensure some edge
conditions are handled properly.

For example, it will now update the state after processing each block
instead of only after processing a block along with any orphans that it
caused to no longer be orphans.

It also now potentially updates the status on every call to the
IsCurrent in order to ensure it is updated properly for callbacks during
block processing.