rpcserver: Add tree param to gettxout.
This adds a required tree parameter to gettxout, since upcoming changes
will be modifying utxo entries to be keyed by an outpoint which includes
the txhash, index, and tree.

For any case where the tree is unknown it will now be expected that the
caller would have to check both trees by calling gettxout for each tree.