blockchain: Support separate test block generation.
This modifies the chain test harness to allow the initial blocks that
are generated to specified heights, such as stake validation height, to
be independently generated without also accepting them to the chain.

It also allows a custom underlying chain generator which houses those
blocks to be provided when creating a test harness.

These two combined provide more flexibility to the tests since they can
choose when (and if) to make the generated blocks available to the
chain.