rpctest: Don't use installed node.
Before these changes rpc tests would use whatever dcrd was installed on
the system. For the run_tests script, this bin was prebuilt. Change
every use of the rpc test harness to use an executable from a
predetermined path, and build a node in a temporary path when not
supplied. Remove building from run_tests.sh.