server: Stop serving v1 cfilters over p2p.
This modifies the server to stop serving v1 cfilters by no longer
responding to the getcfilter, getcfheaders, and getcftypes p2p messages
and no longer advertising support for them via the service flag (even
when indexing is still enabled).

It will still ban peers that attempt to violate the protocol by sending
any of the aforementioned messages after having negotiated a protocol
version high enough that they are knowingly violating it.

This is part of the overall removal of deprecated v1 cfilters.