server/indexers: Remove v1 cfilter indexing support.
This removes support for indexing deprecated v1 cfilters which also
entails removal of the --nocfilters and --dropcfindex flags as well as
all related indexing infrastructure.

The legacy index is now automatically dropped if it exists to ensure
clients that are upgrading are not left with unneeded data in their
databases.

This is part of the overall removal of deprecated v1 cfilters.