blockchain: Whitelist DCP0005 violations.
This updates the block header positional checks to skip the version
check for blocks that are known to violate DCP0005.

Due to a bug that has since been fixed, some blocks were accepted that
violated DCP0005 by still being submitted on an old block version
(version 6 on mainnet) when the majority of the network had already
upgraded to the new version specified by DCP0005 (version 7 on mainnet).

The blocks that violated DCP0005 due to specifying the old version were
otherwise valid and are now whitelisted so that they will be accepted
when fully syncing the chain without checkpoints.