netsync: Move proactive sigcache evict to server.
This moves the proactive signature cache eviction from the sync manager
to the server to more cleanly separate the responsibilities of the
server from the sync manager.  The server controls the lifecycle of the
signature cache and also now handles notifications from the blockchain
when blocks are connected, so the eviction more naturally fits there.

It also removes the SigCache entry from the sync manager Config struct
since it is no longer needed.