blockchain: Fix V3 spend journal migration.
This fixes an issue with the V3 spend journal migration for a case where
the ticket minimal outputs will be incorrectly missing in some migrated
entries.

The issue is that when the utxoset is migrated from V2 to V3, the ticket
minimal outputs are only kept for ticket submission outputs (output 0).
Additionally, the V2 spend journal only stores the ticket minimal
outputs in the last spent entry.  Therefore, for example, in the case
where a ticket has output 0 spent, and output 2 unspent, the ticket
minimal output data will be lost once the utxoset is migrated from V2 to
V3 (since it won't be stored for the unspent output 2, and additionally
it won't be stored for the spent output 0 in the V2 spend journal since
it is not fully spent).

This fixes the issue by migrating the spend journal first using the V2
utxo set where the ticket minimal output data will be available for all
transaction output indexes.