netsync: Remove unused submit block flags param.
This removes the flag from the SubmitBlock method of the sync manager
since those flags are not really intended for general block submission
which is why it is only ever called with blockchain.BFNone.

Also, the intention is to modify the way the sync manager deals with
block submission in the future which this change help simplify.