netsync: Remove orphan block handling.
This removes all orphan block handling from the syncing process and is
part of the overall effort to move to full headers-first syncing as
opposed to the current partial header-first checkpoint-based process.

It should be noted that this change is made in such a way that the
syncing process continues to work properly, however, it is only intended
to be an intermediate step, and, as such, it does result in suboptimal
behavior once it passes the final checkpoint.  That will no longer be
the case one the entire sync process conversion is complete.