netsync: Add header sync progress log.
This modifies netsync to provide progress logging for the initial
headers sync including a progress percentage.

A new method is added that determines the progress of the headers sync
based on an algorithm that calculates the expected number of headers
there should be in the time interval since the current best known header
and the current time given the target block time for the network.  This
is the preferred approach because, unlike the sync height reported by
remote peers, it is difficult to game since it is based on the target
proof of work, but it assumes consistent mining, which is not the case
on all networks, so it is limited to the main and test networks where
that applies.

For the other networks, it falls back to the sync height reported by the
remote peers.

The best known header is now logged at initial load time in blockchain
and the netsync code is updated to make use of the new headers sync
progress logger to tally the number of processed and periodically log
the progress.